Active/disable infinite scroll
* Disable infinite scroll output

The autocomplete component has a property that checks if it should emit
infinite scroll events. This property only gets checked during mounted
and destroyed. If you update the property to false the component should
not emit the infinite-scroll event anymore.

* Add/remove scroll event for Infinitescroll

When checkInfiniteScroll property changes the scroll event should be
added or removed. Just checking the property in the event can cause
an active scroll listener to not get removed on destroy.

* Clean up code

* Add check to see if list exists

Add check to .dropdown-content before attaching event listener

* Remove anonymous event listener

Scroll event was bound using an anonymous function. This way it could
never remove the eventlistener. Removed the parameter from the function
and added it as a variable.

* Clean up code

Co-authored-by: Sjoerd de Wit <sjoerd@superlatief.nl>